Optimized the recipes selelection
Instead of typing the name the use can just select the names

I think we can use this for filter as well
We can just ask the user if they want to filter out certain ingredients
