+++ .idea/workspace.xml
@@ -4,7 +4,14 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="2a44b7c5-26c2-4251-9082-e4c8a18c3f0e" name="Changes" comment="" />
+    <list default="true" id="2a44b7c5-26c2-4251-9082-e4c8a18c3f0e" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/CookingAssistantGUI.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/RecipeManager.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Testing.py" beforeDir="false" afterPath="$PROJECT_DIR$/Testing.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Testing2.py" beforeDir="false" afterPath="$PROJECT_DIR$/Testing2.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Testing3.py" beforeDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
